+++ PicoStreamingAssistantFTUDP/PicoConnector/Pxr.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Runtime.InteropServices;
 
-namespace Pico4SAFTExtTrackingModule;
+namespace Pico4SAFTExtTrackingModule.PicoConnector;
 
 public static class Pxr
 {
